Fix pagination tests
This commit collects the various pagination tests for Organizations back into
the Organizations.test.js file and fixes them so that they pass consistently.
